renamed: config_page
